Python/DBN手写数字识别/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
Pillow
matplotlib